Dropped idea of integrating frontend. Added some more validations and improved previous ones

+++ controllers/products.js
@@ -1,4 +1,9 @@
 const Product = require("../model/product");
+function errorHandler(res,err){
+	console.log("Some error occurred");
+	console.log(err);
+	return res.status(400).json({"msg" : "error occurred"})
+}
 
 async function getAllProducts(req,res){
 	try{
@@ -7,7 +12,8 @@ async function getAllProducts(req,res){
 			console.log({"msg": "no product found"});
 			return res.status(404).json({"msg": "no product found"});
 		}
-		return res.status(200).json({"All Products": products});
+		return res.status(200).json({"All Products": products, "msg": "All products found", "total Products": products.length});
+		// return res.status(200).render("allProducts", {"products": products})
 	}
 	catch(err){
 		console.log("Some error occurred");
@@ -18,9 +24,11 @@ async function getAllProducts(req,res){
 
 async function addProduct(req,res){
 	try{
-		const {pid, name, description, price, inventory} = req.body;
+		const {pid, name, description, price, inventory, category} = req.body;
+		const searchByName = await Product.findOne({name: name});
+		if (searchByName) {return res.status(400).json({"msg": "product with this name already exists"})}
 		const newProduct = await Product.create({
-			pid, name, description, price, inventory
+			pid, name, description, price, inventory, category
 		});
 		if ( !newProduct) {
 			return res.status(400).json({"msg": "err", "newProduct": newProduct})
@@ -29,25 +37,57 @@ async function addProduct(req,res){
 		return res.status(200).json({"msg": "successfully added new product", "product": newProduct});
 	}
 	catch(err){
-		console.log("Some error occurred");
-		console.log(err);
-		res.status(400).json({"msg" : "error occurred"})
+		errorHandler(res,err);
 	}
 }
 
 async function deleteAllProducts(req,res){
 	try{
+		let entries = await Product.find({});
+		if(!entries){return res.status(400).json({"msg" : "error occurred, no entries found"})}
 		await Product.deleteMany({});
-		const item = await Product.findOne({})
-		if (item) {
+		entries = await Product.findOne({})
+		if (entries) {
 			return res.status(400).json({"msg": "err, Entries are still there in DB"})
 		}
 		return res.status(200).json({"msg": "All entries DELETED successfully"});
 	}
 	catch(err){
-		console.log("Some error occurred");
-		console.log(err);
-		res.status(400).json({"msg" : "error occurred"})
+		errorHandler(res,err);
+	}
+}
+async function filterProducts(req,res) {
+	try {
+		const category = req.query.category;
+		if(!category){return res.status(400).json({"msg" : "category is missing"})}
+		if(category === "aromaticPlant"){
+			const list = await Product.find({category: "aromatic plant"});
+			if ( !list) {
+				return res.status(400).json({"msg": "no product found"})
+			}
+			return res.status(200).json({"msg": "list of aromatic plants", "list": list});
+		}if(category === "medicinalPlant"){
+			const list = await Product.find({category: "medicinal plant"});
+			if ( !list) {
+				return res.status(400).json({"msg": "no product found"})
+			}
+			return res.status(200).json({"msg": "list of medicinal plants", "list": list});
+		}if(category === "culinaryHerb"){
+			const list = await Product.find({category: "culinary herb"});
+			if ( !list) {
+				return res.status(400).json({"msg": "no product found"})
+			}
+			return res.status(200).json({"msg": "list of culinary herbs", "list": list});
+		}if(category === "uncategorized"){
+			const list = await Product.find({category: "uncategorized"});
+			if ( !list) {
+				return res.status(400).json({"msg": "no product found"})
+			}
+			return res.status(200).json({"msg": "list of uncategorized plants", "list": list});
+		}
+
+	} catch (err) {
+		errorHandler(res, err);
 	}
 }
 
@@ -55,6 +95,7 @@ async function deleteAllProducts(req,res){
 module.exports = {
 	getAllProducts,
 	addProduct,
-	deleteAllProducts
+	deleteAllProducts,
+	filterProducts
 }
